feat(114-5): CWN equipment extractor + cyberware system_strain (typed float) (#871)
* test: add failing CWN extraction + cyberware system_strain tests (114-5)

RED phase for story 114-5 (ADR-145). Three test files + a synthetic CWN SRD
fixture, all failing because the cwn_equip_extract tool and the CatalogItem
system_strain field do not exist yet:

- tests/cli/test_cwn_equip_extract.py — CWN extraction tool mirroring the WWN
  tool (114-3): verbatim/cwn/wn-free provenance, the net-new Cyberware section
  with a typed system_strain, licensing fail-loud, required --srd-path, and the
  python -m CLI wiring test.
- tests/genre/test_catalog_item_system_strain.py — the schema delta: a typed,
  validated, None-defaulted system_strain int (rejects prose + negative; keeps
  extra=forbid).
- tests/server/test_cwn_inventory_wiring.py — system_strain is a locked
  mechanical field (VerbatimFieldLockError on world re-stat; reskin preserves it)
  and the CWN baseline loads through the real resolve_inventory seam + fires its
  state_transition OTEL span (wiring proof).

ADR-145 D4 settles the licensing fork: CWN is verbatim/wn-free like all four WN
SRDs (the epic's 'verify' parenthetical is resolved by the governing ADR).

* feat(114-5): CWN equipment extractor + float system_strain (cyberware)

Rebuild _parse_cyberware to the real CWN §3.6.7 layout
(Name|Cost|Location|Concealment|System Strain|Effect); System Strain is a
non-negative float (CWN prices common chrome at 0.25/0.5 — an int truncates
0.25 to a 'free' implant). Widen CatalogItem.system_strain, SystemStrainPool,
StrainResult, apply_system_strain, and system_strain_delta_span to float
(max stays int). Per Keith's float ruling, 2026-06-14.


---------

+++ sidequest/game/ruleset/without_number.py
@@ -391,7 +391,7 @@ class WithoutNumberRulesetModule(RulesetModule):
         *,
         core: CreatureCore,
         kind: str,
-        amount: int,
+        amount: float,
         source: str,
         cfg: SwnConfig | None,
         _tracer: trace.Tracer | None = None,
@@ -417,7 +417,9 @@ class WithoutNumberRulesetModule(RulesetModule):
         before = pool.current
         applied = True
         reason = ""
-        requested: int = 0
+        # float: CWN cyberware costs fractional System Strain (0.25/0.5), so a
+        # temporary/permanent install amount can be fractional (Keith, 2026-06-14).
+        requested: float = 0
 
         if kind == "first_aid":
             requested = scfg.first_aid_cost
@@ -651,7 +653,9 @@ class WithoutNumberRulesetModule(RulesetModule):
                     max_spell_level=max_spell_level,
                 )
 
-        return ChargenResources(effort=effort, spellcasting=spellcasting, system_strain=system_strain)
+        return ChargenResources(
+            effort=effort, spellcasting=spellcasting, system_strain=system_strain
+        )
 
     # ------------------------------------------------------------------
     # Chargen contribution methods (ADR-143 Task 10) — background skills +
